Refactor templates, integrate w/ ShrineController
The original templates relied on a ton of variables that have all been
replaced with a player object.

The resurrect messages that were conditional on the type of resurrection
cost have now been made into their own template files and the controller
decides which view part to render based on internal logic.