Add message when using a damaging item on opponent
Resolve #226
The issue was reproducible for shuriken and phosphor powder, both items
that only have a damage component. The conditional branch for items with
damage did not result in a message unless the item was used on self.
This patch adds a message when the item is not used on self.