Replace SESSION with Symfony sessions.
resolve #290
Symfony sessions seem to implement a singleton pattern, except the
constructor returns a single instance of a session PER STORAGE TYPE.
This forces to use Inversion of Control principles to pass the desired
session around. Given our horrible abuse of global functions needing
access to the session, in order to support testing we need a static way
of getting the current session without relying on the conditional
singleton. nw\SessionFactory has been added to make this possible.