Use a variable for phpunit in Makefile
The TEST_RUNNER variable now holds the base invocation for phpunit for
this project, which uses the phpunit installed by composer. Instead of
invoking phpunit directly, we now invoke php directly and use phpunit as
the script file argument. This allows us to load xdebug during
invocation so that xdebug can be removed from the default list of
modules loaded by the environment.