Add support for read-only fields in DAO
There were several ad hoc attempts at getting joins to work in the DAO
for reading that caused problems when trying to save objects because of
the joined in fields. This commit adds the concept of read-only fields
which can be set via a private mutator in the abstract DAO. The DAO then
uses the fields from the VO and removes the read-only fields for the
purposes of writing, eliminating these ad hoc attempts.

Additional simplification comes from removing the overrides for get() in
SkillDAO and PlayerDAO which were unnecessary.