Remove class func from Player, add getMaxHealth
class_identity, class_display_name, and getClassTheme were all being
used to get information about a player's class, most often for display
purposes. Each of these calls made a database call. This is absurd.

Using the updated PlayerVO and PlayerDAO, we now have class identity,
name, and theme in the PlayerVO consistently and reliably. The
controllers and templates that needed these variables are now getting
them directly from the Player object.

getMaxHealth was added as an instance method to remove the
StatsController reliance on a special function with many lines for the
sake of a single value.

Also of note, stats forms and links now use pretty URLs