Refactor health percent calc and render
Health percent rendering required calculating the percentage and passing
it to the view component. This required most controllers to do rendering
math which added complexity. There is now a template function to
calculate a renderable percentage and all it needs is current health and
level. This removes the need for the functions in lib_player and reduces
the complexity of controllers.

Also simplified Player->data()